amazon/requirements.txt
Flask
Flask-SQLAlchemy
Werkzeug
python-dotenv
Flask-WTF
email
openai
